requirements.txt
Jinja2==2.11.2
MarkupSafe==1.1.1
phonenumbers==8.12.6
PyYAML==5.3.1
